+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.6.0 (2024-01-22)
 
 - ENH: Select search order
-- ENH: Report modification date
+- ENH: Report creation date
 
 ## 1.5.1 (2024-01-21)